+++ src/patient_data/tramp.hpp
@@ -16,25 +16,25 @@ public:
     ~Tramp_t();
     void setEnergies();
     void to_file(std::string file, std::string dir = std::string());
-    void scale(float ratio);
+    void scale(double ratio);
     void setWEPLs();
-    std::vector<float> getWEPLs();
+    std::vector<double> getWEPLs();
     void energy_to_internal();
-    void shift_energies(const std::vector<float>& e_, bool units = true);
-    void set_pos(const std::vector< Vector4_t<float> > p);
+    void shift_energies(const std::vector<double>& e_, bool units = true);
+    void set_pos(const std::vector< Vector4_t<double> > p);
     void print(unsigned int n);
     void print(unsigned int n0, unsigned int n1);
     void defaults();
     void read_file_header(std::string f);
 
     std::vector<Spot_t> spots;
-    std::vector<float> energies;
-    std::vector<float> energies_internal;
-    std::vector<float> wepls;
+    std::vector<double> energies;
+    std::vector<double> energies_internal;
+    std::vector<double> wepls;
 
     std::string machine;
 
-    float gigaprotons;
+    double gigaprotons;
     unsigned int nspots;
     std::string patient_id;
     std::string patient_first_name;
@@ -46,8 +46,8 @@ public:
     std::string gantry;
     std::string couch_rotation;
 
-    float z;
-    float zeff;
+    double z;
+    double zeff;
     
 private:
     bool internal_energy_set;
